# lib/api-zod/src/generated/types/arrearsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arrears Management & Budget API
 * OpenAPI spec version: 0.1.0
 */
import type { ArrangementStatus } from "./arrangementStatus";
import type { ArrearsCategory } from "./arrearsCategory";
import type { Frequency } from "./frequency";
import type { RiskLevel } from "./riskLevel";

export interface ArrearsItem {
  id: number;
  creditor: string;
  category: ArrearsCategory;
  balance: number;
  ongoingCharge: number;
  ongoingFrequency: Frequency;
  arrearsPayment: number;
  arrearsFrequency: Frequency;
  riskLevel: RiskLevel;
  status: ArrangementStatus;
  /** @nullable */
  nextReviewDate?: Date | null;
  /** @nullable */
  accountRef?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  objective?: string | null;
  /** @nullable */
  workingPlan?: string | null;
  /** @nullable */
  communicationPosition?: string | null;
  /** @nullable */
  externalAcknowledgement?: string | null;
  /** @nullable */
  externalPaymentIntent?: string | null;
  /** @nullable */
  externalStagedReduction?: string | null;
  /** @nullable */
  externalReviewPoints?: string | null;
  /** @nullable */
  externalChannel?: string | null;
  /** @nullable */
  evidenceLinks?: string | null;
  weeklyOngoing: number;
  weeklyArrears: number;
  weeklyTotal: number;
}