# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arrears Management & Budget API
 * OpenAPI spec version: 0.1.0
 */
import type { Frequency } from "./frequency";

export interface Bill {
  id: number;
  provider: string;
  category: string;
  amount: number;
  frequency: Frequency;
  /** @nullable */
  dueDay?: number | null;
  /** @nullable */
  accountRef?: string | null;
  autopay: boolean;
  weeklyEquivalent: number;
  /** @nullable */
  notes?: string | null;
}